Move SVG inlining script
