+++ src/utils/inline-svgs.mjs
